Add files via upload
Arrays are very helpful for storing lists of data of the same type. This lab will provide experience using arrays of different types including parallel arrays where the data in the 2 arrays is correlated. The lab also deals with presenting output in good form.
Task 1
Create an array to hold 7 numbers with decimal points. Use a loop to read those 7 numbers into the array from the keyboard. Prompt the user as shown. Then, ask the user for a value that you can use to separate large from small numbers. Loop through the array you created and, using the threshold provided by the user, count the number of small numbers (those less than threshold number provided) and large numbers (the others). Report your results.
Task 2
Create a file that contains a number followed by a list of user names (file given for run shown). The number indicates how many names will be found in the file. Your program should read the first number from the file and then create 2 arrays of the corresponding size, one to hold usernames and one to hold passwords. Fill the first array with the names found in the file after the number. Loop through the users array and use the username to prompt the user for the correct password for that username. Store the password in the password array at the same index as the corresponding username. Ex: lcarter is at index 0 in user array and 123qwet would be at index 0 in password array. Now print out all of the names and passwords in 2 columns as shown. You should use tabs to separate the name and password, and the length of the username determines how many tabs you need. Other output should contain some white space as well for readability.
Task 3
Change the password for the third person in the list to only the first 5 characters of the original password. Print that person’s name and new password from the arrays. You can assume that there are at least 3 people in the list and the password was at least 5 characters.